Update utility scripts (use an environment variable to set the root path).

+++ download.tuxfamily.org.push.sh
@@ -1,6 +1,13 @@
 #!/bin/sh
 
-LOCAL_DIR_PATH=~/jdhp/download.tuxfamily.org
+LOCAL_DIR_PATH=${JDHP_ROOT_PATH}/download.tuxfamily.org
 REMOTE_DIR_PATH="tf:~/jdhp/jdhp-repository"
 
-rsync -a -v -e ssh --delete ${LOCAL_DIR_PATH}/ ${REMOTE_DIR_PATH}/
+
+if [ -d ${LOCAL_DIR_PATH} ]
+then
+    rsync -a -v -e ssh --delete ${LOCAL_DIR_PATH}/ ${REMOTE_DIR_PATH}/
+else
+    echo "ERROR: ${LOCAL_DIR_PATH} doesn't exist!"
+fi
+
